conf/machine/marvell-cn10xxx.conf: build 32bit application with 64K alignment
Issue: LIN1024-5227

marvell-cn10xxx uses 64K page size (CONFIG_ARM64_4K_PAGES=y in kernel config).
For 32bit multilib, we'll need to build applications with 64K alignment,
otherwise, we'll see errors like below at runtime:

  root@marvell-cn10xxx:~# /lib/ld-linux-armhf.so.3
  Segmentation fault

Use max-page-size option to fix this issue. Note that this option is added
to both CFLAGS and LDFLAGS for the purpose of covering as many recipes as
possible.

The systemd-boot (bootloader) is an exception. It uses max-page-size=4096
in src/boot/efi/meson.build and it has a hardcoded value in elf2efi.py.

  # EFI mandates 4KiB memory pages.
  SECTION_ALIGNMENT = 4096

We should not change its max-page-size, otherwise we get the following error
when building lib32-systemd-boot:

  raise RuntimeError("ELF segments are not properly aligned.")

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>